CMake, Windows fixes, Fix compilation of conversion test
Added a basic cmake lists script.

Fixed the overload template usage in the conversion test, was causing a compiler error when using AppleClang.

Added header include that shouldn't need to be included . The standard says that including it does nothing but on Windows MSVC when using the header as a standalone file it gave me an issue.